Add packet processor class
This processor can now iterate over a list of files and process the
packets they contain sequentially. This means it is possible to do:

    $ goesproc /path/to/packets/file*

Instead of:

    $ cat /path/to/packets/file* | goesproc

Also, it is no longer possible for a file with a truncated packet to
screw up the entire stream, as in the new mode we ignore incomplete
packets.
